--- additional resources/secrets.py
@@ -1,4 +1,0 @@
-baseURL='http://localhost:8089'
-repository='2'
-user='admin'
-password='admin'
