+++ authenticate.py
@@ -1,5 +1,4 @@
 import requests, json, secrets
-from requests.compat import urljoin, quote
 
 def login():
 	# import secrets
@@ -8,9 +7,7 @@ def login():
 	password = secrets.password
 
 	# attempt to authenticate
-	# following the approach used in ArchivesSnake
-	response = requests.post(urljoin(baseURL, '/users/{user}/login'.format(user=quote(user))),
-		params={"password": password, "expiring": False})
+	response = requests.post(baseURL+'/users/'+user+'/login?password='+password+'&expiring=false')
 
 	if response.status_code != 200:
 		print('Login failed! Check credentials and try again')
