+++ dos.csv
@@ -4,4 +4,4 @@ VIPs like Valerie Addonizio,pic002,path/to/new/jersey/monmouthco/warehouse/pic00
 And Lora Woodford,pic003,definitely/not/on/an/external/harddrive/pic003,8a4eae486afed11aaf4620863ea326f2
 Jumbo the Elephant?,pic004,path/to/archival/storage/pic004,8624a39e58e206cd3b9d45ccb35dfb30
 Handsome Dan?,pic005,path/to/archival/storage/pic005,e14d0f101a0a8a87355eb8b6d95e8b3f
-It's like we're in New England or something…,pic006,path/to/archival/storage/pic006,31803846e42827eed3fcf4392ef0cc5b
+3 Cavaliers,pic006,path/to/archival/storage/pic006,31803846e42827eed3fcf4392ef0cc5b
